requirements.txt
anyio==3.6.2
cffi==1.15.1
click==8.1.3
colorama==0.4.6
cryptography==40.
fastapi==0.95.0
greenlet==2.0.2
h11==0.14.0
idna==3.4
pycparser==2.21
pydantic==1.10.7
PyJWT==2.6.0
PyMySQL==1.0.3
python-decouple==
python-dotenv==1.
sniffio==1.3.0
SQLAlchemy==2.0.8
starlette==0.26.1
typing_extensions
uvicorn==0.21.1